更新了 ios 接口

+++ src/android/hooks/before_plugin_install/unzip_libs.sh
@@ -3,8 +3,5 @@
 cd $(dirname $0)
 cd ../..
 
-if [ ! -e libs ]; then
-    unzip libs.zip
-fi
-
-rm libs.zip
+if [ ! -e libs ]; then unzip libs.zip; fi
+if [ -e libs.zip ]; then rm libs.zip; fi
