+++ src/ios/hooks/before_plugin_install/unzip_libs.sh
@@ -3,8 +3,5 @@
 cd $(dirname $0)
 cd ../..
 
-if [ ! -e TXRTMPSDK.framework ]; then
-    unzip TXRTMPSDK.framework.zip
-fi
-
-rm TXRTMPSDK.framework.zip
+if [ ! -e TXRTMPSDK.framework ]; then unzip TXRTMPSDK.framework.zip; fi
+if [ -e TXRTMPSDK.framework.zip ]; then rm TXRTMPSDK.framework.zip; fi
